0.3.3 method to reset digital offset compensation working; fixed bug in setChannelSources; labview compatibility for method enableRxMessageType

+++ src/messagedispatcher.h
@@ -68,7 +68,7 @@ using namespace e384CommLib;
 
 typedef struct MsgResume {
     uint16_t typeId;
-    uint16_t dataLength;
+    uint32_t dataLength;
     uint32_t startDataPtr;
 } MsgResume_t;
